Actualización Código
Adición código actulizado:
- Controladores
- Servicios
- Repositorios
- Biblioteca Bootstrap

Siguiente: templates, html, adecuar métodos

+++ src/main/java/es/eoi/repository/AlquilerRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AlquilerRepository extends JpaRepository<Alquiler,Long> {
+public interface AlquilerRepository extends JpaRepository<Alquiler,Integer> {
 }
