+++ src/main/java/es/eoi/repository/MaterialReservasRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MaterialReservasRepository extends JpaRepository<MaterialReservas,Long> {
+public interface MaterialReservasRepository extends JpaRepository<MaterialReservas,Integer> {
 }
