+++ src/main/java/es/eoi/repository/ReservasRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReservasRepository extends JpaRepository<Reservas,Long> {
+public interface ReservasRepository extends JpaRepository<Reservas,Integer> {
 }
